Added Prob Evidence BP that adheres to the correct API.
Anytime Prob Evidence BP is not implemented yet as it requires some
careful refactoring since the AnytimeProbEvidence trait conflicts with
AnytimeProbQuery trait (and they both use the same BP class)
