+++ Figaro/src/main/scala/com/cra/figaro/algorithm/ProbEvidenceAlgorithm.scala
@@ -74,10 +74,11 @@ trait ProbEvidenceAlgorithm extends Algorithm {
    * with the initialize method, called when the algorithm starts, and the cleanUp method, called when the algorithm is killed.
    */
   override def initialize(): Unit = {
-    super.initialize()
+    
     savedConditions = for { NamedEvidence(ref, _, _) <- evidence } yield universe.get(ref).allConditions
     savedConstraints = for { NamedEvidence(ref, _, _) <- evidence } yield universe.get(ref).allConstraints
     universe.assertEvidence(evidence)
+    super.initialize()
   }
 
   /**
