+++ Figaro/src/main/scala/com/cra/figaro/algorithm/factored/beliefpropagation/BeliefPropagation.scala
@@ -304,7 +304,7 @@ abstract class ProbQueryBeliefPropagation(override val universe: Universe, targe
   val dependentAlgorithm: (Universe, List[NamedEvidence[_]]) => () => Double,
   depth: Int = Int.MaxValue, upperBounds: Boolean = false)
   extends ProbQueryAlgorithm
-  with ProbabilisticBeliefPropagation with ProbEvidenceBeliefPropagation {
+  with ProbabilisticBeliefPropagation  {
 
   val targetElements = targets.toList
 
@@ -342,74 +342,7 @@ abstract class ProbQueryBeliefPropagation(override val universe: Universe, targe
   }
 }
 
-trait ProbEvidenceBeliefPropagation extends ProbabilisticBeliefPropagation {
 
-  private def logFcn: (Double => Double) = semiring match {
-    case LogSumProductSemiring => (d: Double) => d
-    case SumProductSemiring => (d: Double) => if (d == semiring.zero) Double.NegativeInfinity else math.log(d)
-  }
-  private def probFcn: (Double => Double) = semiring match {
-    case LogSumProductSemiring => (d: Double) => if (d == semiring.zero) 0 else math.exp(d)
-    case SumProductSemiring => (d: Double) => d
-  }
-
-  private def entropy(probFactor: Factor[Double], logFactor: Factor[Double]): Double = {
-    // Even though the variables in each factor are the same, the order of the vars might be different
-    val logFactorMapping = probFactor.variables.map(v => logFactor.variables.indexOf(v))
-    def remap(l: List[Int]) = l.zipWithIndex.map(s => (s._1, logFactorMapping(s._2))).sortBy(_._2).unzip._1
-
-    val e = (0.0 /: probFactor.allIndices)((c: Double, i: List[Int]) => {
-      val p = probFcn(probFactor.get(i))
-      if (p == 0) c else c + p * logFcn(logFactor.get(remap(i)))
-    })
-    e
-  }
-
-  /* Not true mutual information for > 2 factors, but standard for computing Bethe approximation */
-  private def mutualInformation(joint: Factor[Double], marginals: Iterable[Factor[Double]]) = {
-    if (debug) {
-      println(joint.toReadableString)
-      marginals foreach (f => println(f.toReadableString))
-    }
-    val newFactor = (joint /: marginals)((c: Factor[Double], n: Factor[Double]) => c.combination(n, semiring.divide))
-    val mi = (0.0 /: newFactor.allIndices)((c: Double, i: List[Int]) => {
-      val p = probFcn(joint.get(i))
-      if (p == 0) c else c + p * logFcn(newFactor.get(i))
-    })
-    mi
-  }
-
-  /**
-   * Compute the evidence of the model. Returns the probability of evidence on the model. This assumes that BP
-   * has already been run on this algorithm instance.
-   */
-  def computeEvidence(): Double = {
-
-    val factorNodes = factorGraph.getNodes.filter(_.isInstanceOf[FactorNode]).toList
-    val varNodes = factorGraph.getNodes.filter(_.isInstanceOf[VariableNode]).toList
-
-    val nonZeroEvidence = factorNodes.exists(p => beliefMap(p).contents.exists(_._2 != Double.NegativeInfinity))
-
-    if (nonZeroEvidence) {
-      val betheEnergy = -1 * factorNodes.map(f => {
-        entropy(normalize(beliefMap(f)), factorGraph.getFactorForNode(f.asInstanceOf[FactorNode]))
-      }).sum
-      val betheEntropy = {
-        val factorEntropy = -1 * factorNodes.map(f => {
-          entropy(normalize(beliefMap(f)), normalize(beliefMap(f)))
-        }).sum
-        val varEntropy = varNodes.map(v => {
-          (factorGraph.getNeighbors(v).size - 1) * entropy(normalize(beliefMap(v)), normalize(beliefMap(v)))
-        }).sum
-        factorEntropy + varEntropy
-      }
-      math.exp(-1 * (betheEnergy - betheEntropy))
-    } else {
-      0.0
-    }
-  }
-
-}
 
 object BeliefPropagation {
   /**
